moved the search for the pixels to a separate function

+++ cass_pixeldetector/commonmode_calculators.h
@@ -57,7 +57,7 @@ private:
   std::pair<size_t,size_t> _nbrElementsToRemove;
 
   /** mininmal number of pixels that should be present when calculating the mean */
-  size_t _minNbrPixels;
+  int _minNbrPixels;
 };
 
 /** Calculate the common mode by taking the median of pixel values
